Several changes to InvoicePanel behaviour
Selecting an invoice in the Invoice list will now provide the invoice
total in the total textfield.

Upon clicking the 'new invoice' checkbox, the Invoice list will deselect
any selected Invoice, and the list of lineitems will clear. Unchecking
the checkbox before saving the invoice will also clear the lineitem
table.

Invoices, customers, and products added to the database through the
model will be reflected in the InvoicePanel.